requirements.txt
aiohttp
beautifulsoup4
readability-lxml
lxml
